wip! Add create() to widget registry
create() receives a factory and creation options. It adds the registry
provider to the creation options as well as the default store (if any),
if an ID was specified.

Creates the widget and registers it internally, generating a unique
ID if necessary. Returns a promise for a tuple of the ID and widget.

Fixes #22.

TODO:
* Update README
* Tests