Allow initial state to be declared in widget definitions
Like data-state for custom elements, allow initial state to be declared
in widget definitions. If the widget has a store the state is added
before creating the widget. Otherwise it is ignored.

Rejections from add() are ignored. It is assumed these arise when data
already exists in the store. This allows initial state to be provided in
the definition, while applications can still restore previously
persisted state and have that state be used when widgets are rendered.